+++ docs/changelog.md
@@ -1,4 +1,7 @@
 # TBD
+### Features
+* Added a `GetLambdaInfo` endpoint to the API container's API
+* Added a `NetworkContext.getLambdaContext` method for calling the API container method
 
 # 0.8.0
 ### Fixes